Activate cuno individually for each command, so that caches get flushed when commands exit
Also changed instructions to NOT activate a cuno shell before running benchmarks. This is so that our latency hiding is not obscuring the true timing values - if we are already inside a cuno shell, then it's not definite that mkdir, cp uploads, etc. are actually completing when the command exits because cuno has a cache that is tied to the parent shell which still lives